exiftool xslt map: update extension list
This updates exiftool's extension list to include all of the extensions
mentioned in the exiftool 9.05 documentation.

A few of these may in the wrong category; for instance, I wasn't sure
if the metadata in an InDesign document is more like an "image" or a
"document".

There are a few false categorizations here because of the reliance on
extension to determine type. Some extensions, e.g. XMP and OGG, could
contain audio, video, or both. In the future mapping should probably be
done by MIMETYPE, not extension.

For ease of maintenance, this also alphabetizes the list of extensions
per section.