exiftool: add video metadata mapping fields
These are by no means exhaustive, but this does add a number of the
common metadata fields which can be found in video.

The new field names shouldn't be considered set in stone until an
official FITS release so feel free to change them up if you don't like
what I went with.

A thought: so many field names are in common with images, does it make
sense to process images and videos with separate XSLT logic?